feat: add activity message for SLA
